add some framework for simulating the SCE
"almost" all open source, there is an ahb/axi adapter that i'm
cheating on right now and relying on an outside repo to get
the sim to run. This is a pretty easy/common part to swap out
for a full OSS equivalent later on, now focused on just getting
the sim up and running.